+++ src/components/checkbox/index.ts
@@ -1,0 +1,2 @@
+export * from "./Checkbox"
+export type { CheckboxProps, CheckboxVariant } from "./Checkbox.model"